fix(polling): refreshing of user accessTokens
by requiring an accessToken to have an `expires` timestamp, which is
utilized by the [spotify sdk] to determine if a new accessToken needs to
be requested with the `refresh_token`.

As the accessToken now contains a couple of fields, it was addtionally
extracted into its own table. This allows checking the db for token
updates by the `updatedAt` timestamp.

[spotify sdk]: https://github.com/spotify/spotify-web-api-ts-sdk/blob/main/src/auth/ProvidedAccessTokenStrategy.ts#L38-L45